Implement more code cleanup
